PB-579: Fix pre-release tagging issue
We could end up on a tag already exists with pre-release tag. The issue occurred
when a minor pre-release tag was created followed by a patch pre-release tag,
then the next pre-release tag was the same as the last patch. For example
such tag history would have led to an issue

 - v1.11.1-beta.1
 - v1.12.0-beta.1
 - v1.11.0
 - v1.10.0
 - ....

In this case the the next pre-release tag would have been v1.11.1-beta.1 instead
of v1.12.0-beta.2

The issue was that pre-release bumping logic was not done correctly as the tags
are sorted by semver and not how they are added. In the first place above
creating a v1.11.1-beta.1 after v1.12.0-beta.1 was already not correct and
should not have happened. This was du to the fact that the first pre-release
tag was done using the default bump type of minor and that the second pre-release
tag was created with a #patch in a commit message which should have resulted
to v1.12.1-beta.1 or v1.12.0-beta.2 and not v1.11.1-beta.1.

Now if a pre-release tag already exist we bump it (ignoring the bump type) unless
the bump type is higher as the previous pre-release. This should add a more
consistent way of tagging.